Add confirmation on deletion.

+++ android/app/src/main/java/org/darisadesigns/polyglotlina/android/ui/Lexicon/LexemeInfoActivity.java
@@ -1,5 +1,6 @@
 package org.darisadesigns.polyglotlina.android.ui.Lexicon;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 
@@ -20,6 +21,7 @@ import android.widget.Toast;
 
 import org.darisadesigns.polyglotlina.DictCore;
 import org.darisadesigns.polyglotlina.Nodes.ConWord;
+import org.darisadesigns.polyglotlina.android.AndroidInfoBox;
 import org.darisadesigns.polyglotlina.android.AndroidPropertiesManager;
 import org.darisadesigns.polyglotlina.android.PolyGlot;
 import org.darisadesigns.polyglotlina.android.R;
@@ -113,16 +115,27 @@ public class LexemeInfoActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case R.id.action_delete:
-                try {
-                    core.getWordCollection().deleteNodeById(conWord.getId());
-                    finish();
-                } catch (Exception e) {
-                    core.getOSHandler().getIOHandler().writeErrorLog(e);
-                    core.getOSHandler().getInfoBox().error(
-                            "Deletion Error",
-                            "Unable to delete word: " + e.getLocalizedMessage()
-                    );
-                }
+                ((AndroidInfoBox)core.getOSHandler().getInfoBox()).yesNoCancel(
+                        "Are you sure?",
+                        "Do you want to delete this lexeme?\nThis action can't be undone.",
+                        this,
+                        (dialog, which) -> {
+                            if (which != DialogInterface.BUTTON_POSITIVE) {
+                                return;
+                            }
+                            try {
+                                core.getWordCollection().deleteNodeById(conWord.getId());
+                                finish();
+                            } catch (Exception e) {
+                                core.getOSHandler().getIOHandler().writeErrorLog(e);
+                                ((AndroidInfoBox)core.getOSHandler().getInfoBox()).error(
+                                        "Deletion Error",
+                                        "Unable to delete word: " + e.getLocalizedMessage(),
+                                        this
+                                );
+                            }
+                        }
+                );
                 return true;
             case android.R.id.home:
                 onBackPressed();
@@ -147,9 +160,10 @@ public class LexemeInfoActivity extends AppCompatActivity {
                     core.getWordCollection().deleteNodeById(conWord.getId());
                 } catch (Exception e) {
                     core.getOSHandler().getIOHandler().writeErrorLog(e);
-                    core.getOSHandler().getInfoBox().error(
+                    ((AndroidInfoBox)core.getOSHandler().getInfoBox()).error(
                             "Deletion Error",
-                            "Unable to delete word: " + e.getLocalizedMessage()
+                            "Unable to delete word: " + e.getLocalizedMessage(),
+                            this
                     );
                 }
             }
